# docs/openapi.yaml
openapi: 3.0.0
info:
  title: AgentFlow API
  description: |
    API documentation for the AgentFlow application.
    This API provides endpoints for task management, workflow execution, and authentication.
  version: 1.0.0
  contact:
    name: AgentFlow Support
    email: support@agentflow.com
servers:
  - url: /api
    description: Default API server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      required:
        - error
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        taskType:
          type: string
          description: Type of the task
          example: "fetchCRMReport"
        taskText:
          type: string
          description: Original task text
          example: "Fetch yesterday's sales report from VinSolutions for dealer ABC123"
        taskData:
          type: object
          description: Task-specific parameters
          example:
            platform: "vinsolutions"
            dealerId: "ABC123"
            reportType: "sales"
            dateRange: "yesterday"
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current status of the task
          example: "completed"
        userId:
          type: string
          description: ID of the user who created the task
          example: "user_123456"
        result:
          type: object
          description: Result of the task execution
          example:
            title: "Sales Report Summary"
            insights: ["Total sales increased by 15% compared to last week", "SUV sales are trending upward"]
            actionItems: ["Follow up with potential customers", "Focus marketing on SUV models"]
        error:
          type: string
          description: Error message if the task failed
          example: "Failed to authenticate with VinSolutions CRM"
        createdAt:
          type: string
          format: date-time
          description: When the task was created
          example: "2023-05-15T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the task was last updated
          example: "2023-05-15T14:35:22Z"
      required:
        - id
        - taskType
        - taskText
        - status
...
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the job
          example: "job_7f8d9e6a5b4c3d2e1f"
        taskId:
          type: string
          format: uuid
          description: ID of the associated task
          example: "task_1a2b3c4d5e6f7g8h9i"
        status:
          type: string
          enum: [pending, running, completed, failed]
          description: Current status of the job
          example: "completed"
        priority:
          type: integer
          description: Job priority (higher number means higher priority)
          example: 5
        result:
          type: object
          description: Result of the job execution
          example:
            downloadedFile: "/tmp/reports/sales_report_20230515.csv"
            recordCount: 156
            processedAt: "2023-05-15T14:32:45Z"
        error:
          type: string
          description: Error message if the job failed
          example: "Timeout while connecting to CRM server"
        attempts:
          type: integer
          description: Number of attempts made to execute the job
          example: 2
        maxAttempts:
          type: integer
          description: Maximum number of attempts allowed
          example: 3
        startedAt:
          type: string
          format: date-time
          description: When the job started execution
          example: "2023-05-15T14:30:15Z"
        completedAt:
          type: string
          format: date-time
          description: When the job completed execution
          example: "2023-05-15T14:33:22Z"
        createdAt:
          type: string
          format: date-time
          description: When the job was created
          example: "2023-05-15T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the job was last updated
          example: "2023-05-15T14:33:22Z"
      required:
        - id
        - taskId
        - status
        - priority
...
paths:
...
  /tasks:
    get:
      summary: List all tasks
      description: Get a list of all tasks
      operationId: listTasks
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new task
      description: Submit a new task for processing
      operationId: createTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  description: Task description
              required:
                - task
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  jobId:
                    type: string
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tasks/{taskId}:
    get:
      summary: Get task by ID
      description: Get details of a specific task
      operationId: getTask
      security:
        - cookieAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /test-parser:
    post:
      summary: Test task parser
      description: Test the task parser with a given input
      operationId: testParser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  description: Task description to parse
              required:
                - task
      responses:
        '200':
          description: Parsed task result
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    $ref: '#/components/schemas/Task'
                  executionPlan:
                    type: object
                    description: Execution plan for the task
                  error:
                    type: string
                    description: Error message if parsing failed
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /submit-task:
    post:
      summary: Submit task for immediate execution
      description: Submit a task for immediate processing with high priority
      operationId: submitTask
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  description: Task description
              required:
                - task
      responses:
        '201':
          description: Task submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  jobId:
                    type: string
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /jobs:
    get:
      summary: Get all jobs
      description: Returns a list of all jobs with optional filtering by status
      operationId: listJobs
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          description: Filter jobs by status
          schema:
            type: string
            enum: [pending, running, completed, failed]
        - name: limit
          in: query
          description: Maximum number of jobs to return
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /jobs/enqueue/{taskId}:
    post:
      summary: Enqueue a new job for a task
      description: Manually enqueue a new job for a task
      operationId: enqueueJob
      security:
        - cookieAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          description: Task ID
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                priority:
                  type: integer
                  default: 1
                  description: Job priority (higher number means higher priority)
      responses:
        '200':
          description: Job enqueued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Job enqueued successfully
                  jobId:
                    type: string
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...